requirements.txt
dbt-core==1.5.2
dbt-snowflake==1.5.2
apache-airflow-providers-dbt-cloud
apache-airflow-providers-snowflake==5.8.0
snowflake-connector-python
werkzeug==2.3.7